Refactor API calls
It was getting hard to manage the various API calls because they were all getting loaded into index.js. Instead, they are now separated out into their respective modules under src/calls. Each module exports a function that appends the various API calls to the main object, so that we maintain the singleton but now we group them by functionality.

This also makes significant improvements to the call structure by standardizing under one api-call module. Everything that hits the API goes through that. As a consequence and simplification, we don't explicitly check signed in status for each API call. This makes the wrapper much simpler and easier to maintain.